+++ apps/ui/src/components/Nav/desktop/NavCategoryComponent.ts
@@ -1,8 +1,8 @@
-import type { Component } from '../../types/component.ts';
-import { BaseComponent } from '../BaseComponent/BaseComponent.ts';
-import { Typography } from '../Typography/Typography.ts';
+import type { Component } from '../../../types/component.ts';
+import { BaseComponent } from '../../BaseComponent/BaseComponent.ts';
+import { Typography } from '../../Typography/Typography.ts';
 import { NavCategoryItemComponent } from './NavCategoryItemComponent.ts';
-import type { NavCategory } from './types.ts';
+import type { NavCategory } from '../types.ts';
 
 export class NavCategoryComponent extends BaseComponent {
   #header: Component;
