+++ apps/ui/src/components/Nav/desktop/NavCategoryItemComponent.ts
@@ -1,6 +1,6 @@
-import { BaseComponent } from '../BaseComponent/BaseComponent.ts';
-import { Typography } from '../Typography/Typography.ts';
-import type { NavCategoryItem } from './types.ts';
+import { BaseComponent } from '../../BaseComponent/BaseComponent.ts';
+import { Typography } from '../../Typography/Typography.ts';
+import type { NavCategoryItem } from '../types.ts';
 
 export class NavCategoryItemComponent extends BaseComponent {
   #link: BaseComponent;
