+++ apps/ui/src/components/Nav/desktop/desktopNav.css
@@ -1,4 +1,4 @@
-.Nav-root {
+.DesktopNav-root {
   border-right: 1px solid var(--color-border);
   display: flex;
   flex-direction: column;
@@ -38,7 +38,7 @@
 }
 
 @media (max-width: 600px) {
-  .Nav-root {
+  .DesktopNav-root {
     display: none;
   }
 }
